backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python==4.11.0.86
numpy==2.2.4
pydantic==2.11.0
sleap-io==0.6.5
requests==2.32.3
scipy>=1.12.0
scikit-learn>=1.3
statsmodels>=0.14
umap-learn>=0.5
reportlab>=4.0
pillow>=10.0
